upper/requirements.txt
numpy
opencv_python
Pillow
PyBluez
pyserial
torch
torchvision
tqdm
